Add a non zero exit code on connection failures
Very very minor thing, but it bit me when I was putting some error handling around it. Prior it'd just `exit()` and thus be zero, rather than a usable one to catch the error.

Went with 110 which if I recall correct is connection timed out, seemed reasonable enough to use that.